requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.1
nest-asyncio==1.6.0
pymongo==4.11.3
python-telegram-bot==22.0
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.13.0
urllib3==2.3.0
